Harden System Security v1.0.13.0 | AppControl Manager v2.0.43.0 (#866)
What's New

For Harden System Security and AppControl Manager

Removed 2 dependencies, CommunityToolkit.WinUI.Behaviors and Microsoft.Xaml.Behaviors.WinUI.Managed. The result is reduced package size by ~150 KB for each app bundle and improved overall performance since the new codes are more targeted and optimized. This repository always strives for leaner and more efficient codebases and less dependency usage.

Fixed a visual issue in Remote Desktop sessions on hosts without a GPU where the new home page's stage background displayed in a too bright color, making the text hard to read.

Bumped app versions.

Updated and localized the Home page's content.

You can now use the keyword left and right arrow keys to navigate the carousel in the Home page. As a reminder, you can also use touch gestures to navigate it.


For Harden System Security Only

Added an extra check before applying/removing/verifying some of the security measures in the Microsoft Defender category related to the Defender antivirus, to make sure they have been initialized by the OS first. Without this check, you would see an error in the UI, but from now on they will be logged and gracefully skipped instead. These security measures are rather new compared to others so might not be available immediately after clean installing Windows OS or setting up a new VM, they will become automatically available after the OS is updated or restarted. The reason for this change is to improve the user experience and avoid confusion because an error could be perceived as a bug in the app or a failure of the operation, while in fact it is just a timing issue and expected.

Improved the user experience: When an operation is in progress using the Apply All, Remove All, or Verify All buttons on a page, the other buttons are temporarily disabled until it completes. The active button remains enabled, allowing the operation to be canceled. Similarly, when the Apply, Remove, or Verify button of an individual security measure is running, all other buttons are disabled until it finishes. This change conveys the intended usage of the UI elements more clearly.

Changed behavior: When you press the Apply All button of each page/category, items that belong to Sub-Categories are no longer applied. In order to apply them, you will have to either press the Apply button on each of them individually or select them and then use the Apply Selected button. This change is made to prevent unintended application of sub-categories that usually need more attention or user awareness and confirmation before being applied.

Added new tooltips and help texts when you hover over these buttons to explain their behavior.
The Status Overview section in each page now has checkboxes next to each status item, you can use these new UI elements to filter the security measures based on their current status, such as Applied, Not Applied or Undetermined (N/A). Additionally, the exact count of security measures for each status is displayed next to the corresponding status item.

Improved localization.

Fixed an issue with verifying multi-line registry keys for security policies whose desired value is empty. When using verify button for them, they would show as Not Applied even if they were Applied (they were in fact correctly empty). It's been fixed now. The issue was only isolated for verification, applying and removing them worked correctly.


For AppControl Manager Only

The View File Certificates page now supports drag and drop when the app is running unelevated. Simply drag and drop your files on the page to view their certificates.