AppControl Manager v2.0.45.0 (#881)
Added a new button to the sidebar that allows you to optimize memory usage of the app and try to reduce it. AppControl Manager is already highly optimized but this gives users more control over its memory usage. The app will generate a detailed report in the Logs page that you can check out, showing the type and amount of memory changes.

The same feature has been added for Harden System Security app as well.
In Scan results, for the Signing Scenario column, the status is now displayed as "UserMode" or "KernelMode". Previously, they would show up as 0 and 1, indicating Kernel Mode and User mode respectively. This change improves readability, user experience and addresses the following community feedback.

Updated dependencies to the latest versions.

The View File Certificates page now displays a lot more information about each signer and has the ability to export the results to JSON for backup. It also features a new button that displays CMS details of the selected signed file if present.

The System Information page now lets you easily open the selected deployed policy in Policy Editor where you can view all of its details, modify it and save it to a file. This is very useful to know what rules are included in the deployed policies on your system. It also supports protected System policies that come by default with the OS.