Update Harden-Windows-Security.ps1
Removed the extra checks and actions related to Attack Surface reduction exclusions or registry checks during AppControl Manager installation via the script. If user needs to install it in an environment with ASR rules applied, they should use the Microsoft Store version instead which is signed and compatible with ASR rules.