Harden System Security v1.0.20.0 (#906)
What's New
Updated dependencies to the latest versions.

Added a new dropdown button to the Sidebar, responsible for enabling or disabling Nested Virtualization for all Hyper-V virtual machines on the system. This feature allows virtual machines to run other virtual machines within them or utilize OS security features that rely on virtualization, enhancing testing and development capabilities. Note that all VMs need to be turned off for this operation to succeed.

The Sidebar's pane is now resizable and no longer has a fixed width.

Improved localizations.

Added a new dropdown button to the Sidebar, responsible for enabling/activating or disabling/deactivating the Ultimate Performance power plan on the system. This power plan is designed to provide the highest level of performance for demanding workloads by minimizing power-saving features. It is particularly beneficial for high-performance computing tasks, gaming, and other resource-intensive applications.

Added full German language localization resources for the Harden System Security app to the repository. However, it will be available in the GUI and Microsoft Store in the next version because it needs to go through reviews first.

Added CLI support for applying and removing the security measures based on Presets in the Protect page.

Added CLI support for applying the security measures based on device usage intents in the Protect page.

Full console support is provided and the documentation has been updated accordingly. You can now use the designated features of the application entirely from the command line interface (CLI) without needing to interact with the graphical user interface (GUI). This is particularly useful for advanced users and system administrators who prefer command-line operations for automation and scripting purposes.

It completes the following feature request.