Harden-System-Security-v1.0.23.0 (#925)
What's New
Updated dependencies to the latest version.

Removed any dependency on secedit that used to exist; From now on, all security policies are applied, verified and/or removed by the Harden System Security app itself via low level APIs.

Changed the device usage intent for Administrator Protection policy in the User Account Control category to only apply to School, Business, Specialized and Enterprise devices, excluding Development and Gaming device usage intents.

Added links in the UI to many security measures that didn't have them.

Added an Export button to the Optional Windows Features page. Use it to export the list of all optional Windows features and capabilities along with their current status and details to a structured JSON file.

Added an Export button to the Attack Surface Reduction rules page. Use it to export the list of all Attack Surface Reduction rules and their status to a structured JSON file.

Added an Export button to every page that offers security measures configurations. You'll be able to export all of the data about your system for that specific category quickly to a JSON for safe keeping or tracking/backup.

Added an Export button to the Manage Installed Apps page. Use it to export the list of all installed apps along with their full details to a structured JSON file.

With this update, every feature in the Harden System Security app offers a way to export the data that you see in it. Some pages such as Audit Policies allow you to modify the exported data and import them back to apply them again on the system. In future updates, this behavior will be expanded to other features as well.