Harden-System-Security-v1.0.25.0
The Harden System Security app now provides comprehensive system report generation capabilities, allowing users to gain deeper insights into their system's security status and save the results to a file for further analysis. All of the security categories participate in this report, including Microsoft Security baselines and Microsoft 365 apps security baselines.

The name of the device, user account and time is also mentioned in the report header so you can easily identify when and where the report was generated.

The report provides a security score, the exact number of items in each category and the total number of compliant and non-compliant items.

After you generate a system report, whether it's a full system report that includes all categories or only contains 1 category, you can import and apply it to any system where Harden System Security app is installed.

There are 2 modes of application: Partial and Full.

In partial mode, only items marked as `Applied` in the report will be applied to the system.

In full mode, the system will be fully synchronized according to the report, that means any security measure marked as `Applied` will be applied onto the system and any security measure marked as `Not Applied` will be removed from the system.

Let's say you've configured a system to your liking and it's in a compliant golden state, you can back up the categories you want or the entire system state to a JSON file and import it on as many workstations as you want.

**The great thing about all of this? You can totally automate it, thanks to the new CLI support. [Documentation for CLI usage available here](https://github.com/HotCakeX/Harden-Windows-Security/wiki/Harden-System-Security#commandline-interface-cli-support).**

> [!NOTE]\
> In this update, restoration does not apply to the following categories:
> * The 2 Microsoft Security baseline categories.
> * Attack Surface Reduction rules category.
>
> Rest assured that these 3 categories will be supported for restoration in the next update.

Completely reworked the design and implementation of the search experience in the Harden System Security app. The top search bar now serves as the unified search bar, offering localized contextual search results across all categories and pages.

For example, if you're looking for enabling or disabling "Administrator Protection", you no longer have to know the category it is on beforehand or browse the categories one by one to find it, you can simply type a few letters and the search bar will automatically show you accurate suggestions.

The search results being localized mean if you're using the app in German or Polish for instance, then you can type using your native language in the search bar and the results will be shown accordingly. The icon of the page is also displayed for each suggestion, helping with easy selection of the page.

* Completed this feature request: https://github.com/HotCakeX/Harden-Windows-Security/issues/657

* Completed another feature request: https://github.com/HotCakeX/Harden-Windows-Security/issues/854

* Updated dependencies to the latest versions.

<br>